Add console error message for pam-ipmicheck
When changing the password of an IPMI user with passwd command, if a
invalid password is provided, it only gives the default error message.
This patch adds a detailed error message for interactive console in
pam-ipmicheck.

Tested:
Set a 22-char password for an ipmi user with passwd command, it prints:
"Username 4 / Password 22 exceeds IPMI 16/20 limit"

Change-Id: I1267939ae3cdee81e3efad7d6ec45f572bd83f3d
Signed-off-by: Jiaqing Zhao <jiaqing.zhao@intel.com>